packages/infrastructure/pyproject.toml
[project]
name = "lintel-infrastructure"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "lintel-contracts",
    "lintel-domain",
    "structlog>=24.4",
    "asyncpg>=0.30",
    "sqlalchemy[asyncio]>=2.0",
    "presidio-analyzer>=2.2",
    "presidio-anonymizer>=2.2",
    "litellm>=1.55",
    "slack-bolt>=1.21",
    "slack-sdk>=3.33",
    "pydantic>=2.10",
    "pydantic-settings>=2.7",
    "httpx>=0.28",
    "tenacity>=9.0",
    "opentelemetry-sdk>=1.29",
    "opentelemetry-exporter-otlp>=1.29",
    "cryptography>=44.0",
    "nats-py>=2.9",
    "boto3>=1.35",
    "fastapi-mcp>=0.4.0",
]

[tool.uv.sources]
lintel-contracts = { workspace = true }
lintel-domain = { workspace = true }

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
only-include = ["src/lintel/infrastructure"]

[tool.hatch.build.targets.wheel.sources]
"src" = ""
